staging: dpaa2-ethsw: setup the STP state for all installed VLANs
Setup the STP state for all VLANs installed on the port. This is also
avoiding the error situation when the DEFAULT_VLAN_ID is not installed
on the port (thus the firmware complains that it cannot setup the
required STP state).

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Link: https://lore.kernel.org/r/20200721091919.20394-4-ioana.ciornei@nxp.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 4acfc4393e90e4dfa6ebc6e43545296ca1d928fe)
Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>