driver: mtd: spi-nor: fsl-quadspi: Flush any pending data on suspend call
Flush both RX, TX FIFO even if there is pending transaction in only one of
them.

LPM20 was passing/suceeding after addition of this step in suspend call of
fsl-quadspi.c. Asleep LED was asserted and system did wake up from sleep on
expiry of time.

Following command was used for testing
 echo '7       4       1       7' > /proc/sys/kernel/printk
 echo 0 > /sys/class/rtc/rtc1/wakealarm
 echo +10 > /sys/class/rtc/rtc1/wakealarm

 if [ $? -eq 0 ] ;then
 echo mem > /sys/power/state
 fi

This Patch is only required for LTS4.19 as fsl-quadspi.c is re-implemented
in Kernel-5.4 due to introduction of new SPI-MEM frame. Pending data is
not observed with new implementation in kernel-5.4

Cc: stable@vger.kernel.org # v4.19
Signed-off-by: Ran Wang <Ran.Wang_1@nxp.com>
Signed-off-by: Ashish Kumar <Ashish.Kumar@nxp.com>