dpaa2-eth: properly take into account the _NO_MAC_FILTER option
When the DPNI_OPT_NO_MAC_FILTER option is set on a specific DPNI, the
corresponding network interface is not provisioned with a MAC filtering
table. In this case, do no try to explicitly add the broadcast address
to the filtering list since it would lead to a definitive error.
Also, when the option is encountered the interface should be in
promiscous mode.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>