dpaa2-eth: do not call link_state_update() on an _open
With the addition of the MAC/PHY support directly in the dpaa2-eth
driver the dpmac object is started on ndo_open() and not at system boot
thus the dpmac cannot process the link-up before we send it. Remove the
link_state_update() call because there is no point in it anymore.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>