staging: dpaa2-mac: fix link state handling
Because the MC firmware enables the switch interfaces at create time
this leads to several problems targeting the PHY start/stop sequences.
This in turn leads to a behavior where the link on the PHY is up but the
firmware is not notified of it so it still sees the ports as down.

First, make the same check as the PHY library before actually calling
phy_start/phy_stop so that we do not end up with a WARNING log.
Secondly, if the PHY was already started but the firmware requests a
LINK_UP then restart the aneg on the PHY so that we trigger another
link-up interrupt from the PHY library. Essentially, this check
determines the moment when the two states (firmware and Linux) are not
in sync and tries to sync them again.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>