staging: dpaa2-mac: fix race between PHY and MC
When a pause frame configuration change happens on a dpdmux port, the MC
firmware will send both the LINK_DOWN_REQ and LINK_UP_REQ interrupts
accompanied by the new pause frame configuration. If the
dpmac_get_link_cfg() command is delayed until after the phy_stop() call
is made this leads to the fact that the requested configuration is
overwritten by the PHY irq callback.
Also, keep track of the autoneg state and when the autoneg is disable
the only thing that should change is the link up (or not) field.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>