staging: dpaa2-ethsw: add a dummy ACL table
Proper broadcast functionality is only available when an ACL table is
created and all switch interfaces are added to it. This is due to the
current implementation of the broadcast isolation feature which relies
on a user created ACL table in which it adds an extra entry for
broadcast traffic.
Patch the dpaa2-ethsw driver in order to keep broadcast functionality
working.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>