net: enetc: honor the min_frag_size frame preemption parameter
Using our downstream ethtool patch, it is possible to specify a
minFragSize value:

ethtool --set-frame-preemption eno0 fp on preemptible-queues-mask 0x5e min-frag-size 124

Currently the driver ignores the min-frag-size option and keeps it at
60, the hardware default value.

In hardware there are two fields in the Merge Control and Status
Register (MAC_MERGE_MMCSR):

LAFS (Local additional fragment size) holds the minFragSize required by
our RX side. It is not used by hardware and just some storage in case
the LLDP daemon wants to create an Additional Ethernet Capabilities TLV.
Frankly this is useless.

RAFS (Remote additional fragment size) holds the minFragSize required by
the RX of the link partner. Hardware uses this to adjust the minFragSize
on TX. This is what needs to be programmed.

We also change the .get_preempt method to retrieve the RAFS register.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>