dtb : la1238-rdb eMMC clock tunning and false data interrupt workaround
la1238-rdb board sometimes fails to tune clock for eMMC in HS400 mode.
This lead to failure of auto mounting of eMMC partations and false data
interrupt.
eMMC is found to be working reliable and as per specs in HS200  mode.
This workaround remove HS400 mode support and highest supported mode is
HS200.

Signed-off-by: Pramod Kumar <pramod.kumar_1@nxp.com>