staging: dpaa2-switch: create and assign an ACL table per port at probe time
At probe time, create for each switch port a private ACL table. At the
moment, no rules are being added in the ACL tables, this will comes in
the next patches.

This patch is a squash and backport of the following two patches which
can be found in upstream Linux:

commit 90f071023529 ("dpaa2-switch: create and assign an ACL table per
port")
commit 1b0f14b6c218 ("dpaa2-switch: create a central
dpaa2_switch_acl_tbl structure")

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>