dpaa2-switch: add debug capabilities for ACL rule management
Add debugging prints for ACL rule management. If enabled, the prints
will show which ACL table is being used by a specific port, what rules
are being added into an ACL table and list the entire content of the
key.
This is helpful when debugging and creating an initial setup using ACL
rules.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>