net: enetc: fix timestamp handling in receive path with XDP_PASS
The XDP_PASS path in the driver is missing the SKBTX_HW_TSTAMP_NETDEV
flag when building the skb for the regular network stack path. This
causes sending bogus timestamps as the driver is enabling free running
support by default [1] and without this flag, the timestamp conversion is
not enabled in the core socket code.
To future-prrof this, move the flag setting to a more centralized
function and put it together with the hwtimestamp code.

Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>