net: dsa: netc: restore hsr config when switch resumes
If NETC is powered off when system suspends, configuration of HSR will
be lost. Restore the configuration when system resumes.

In order to restore the HSR configuration, We store the HSR type and
ISGT_EID of each switch port. Clean old ISGT_EID in software and set HSR
on each port by using HSR configuration.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>